push.
Figure out a plan to incorporate kitaes model to the streamlit. right now my sentiment analysis outputs the 1) date 2) and vector sentiment for every topic. i think this is the input to the .pkl file
Then output:
the VIX number +- confidence interval. Output this in chart too, along with the historical VIX.
descriptive correlation of FOMC to VIX (SHAP, LIME on which topics and why.
do backtest on what happened in 2008, 2020. which words or topics are most relevant. see if this makes sense, to confirm out confidence